[Alert] Move from lab to core
