+++ docs/src/pages/components/alert/ColorAlerts.js
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import Alert from '@material-ui/lab/Alert';
+import Alert from '@material-ui/core/Alert';
 
 export default function ColorAlerts() {
   return (
