+++ docs/src/pages/components/alert/OutlinedAlerts.js
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import Alert from '@material-ui/lab/Alert';
+import Alert from '@material-ui/core/Alert';
 
 const useStyles = makeStyles((theme) => ({
   root: {
