+++ docs/src/pages/components/snackbars/CustomizedSnackbars.js
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import Button from '@material-ui/core/Button';
 import Snackbar from '@material-ui/core/Snackbar';
-import MuiAlert from '@material-ui/lab/Alert';
+import MuiAlert from '@material-ui/core/Alert';
 import { makeStyles } from '@material-ui/core/styles';
 
 const Alert = React.forwardRef(function Alert(props, ref) {
