+++ packages/material-ui-lab/src/index.js
@@ -1,10 +1,4 @@
 /* eslint-disable import/export */
-export { default as Alert } from './Alert';
-export * from './Alert';
-
-export { default as AlertTitle } from './AlertTitle';
-export * from './AlertTitle';
-
 export { default as Autocomplete } from './Autocomplete';
 export * from './Autocomplete';
