+++ packages/material-ui/src/Alert/Alert.js
@@ -1,10 +1,12 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import clsx from 'clsx';
-import { withStyles, lighten, darken, useThemeVariants } from '@material-ui/core/styles';
-import Paper from '@material-ui/core/Paper';
-import IconButton from '@material-ui/core/IconButton';
-import { capitalize } from '@material-ui/core/utils';
+import { useThemeVariants } from '@material-ui/styles';
+import withStyles from '../styles/withStyles';
+import { darken, lighten } from '../styles/colorManipulator';
+import capitalize from '../utils/capitalize';
+import Paper from '../Paper';
+import IconButton from '../IconButton';
 import SuccessOutlinedIcon from '../internal/svg-icons/SuccessOutlined';
 import ReportProblemOutlinedIcon from '../internal/svg-icons/ReportProblemOutlined';
 import ErrorOutlineIcon from '../internal/svg-icons/ErrorOutline';
