+++ packages/material-ui/src/AlertTitle/AlertTitle.d.ts
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { InternalStandardProps as StandardProps } from '@material-ui/core';
+import { InternalStandardProps as StandardProps } from '..';
 
 export interface AlertTitleProps extends StandardProps<React.HTMLAttributes<HTMLDivElement>> {
   /**
