+++ packages/material-ui/src/AlertTitle/AlertTitle.js
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
-import { withStyles } from '@material-ui/core/styles';
-import { Typography } from '@material-ui/core';
 import clsx from 'clsx';
+import withStyles from '../styles/withStyles';
+import Typography from '../Typography';
 
 export const styles = (theme) => ({
   /* Styles applied to the root element. */
