+++ packages/material-ui/src/index.d.ts
@@ -92,6 +92,12 @@ export * from './AccordionDetails';
 export { default as AccordionSummary } from './AccordionSummary';
 export * from './AccordionSummary';
 
+export { default as Alert } from './Alert';
+export * from './Alert';
+
+export { default as AlertTitle } from './AlertTitle';
+export * from './AlertTitle';
+
 export { default as AppBar } from './AppBar';
 export * from './AppBar';
