+++ packages/material-ui/src/index.js
@@ -18,6 +18,12 @@ export * from './AccordionDetails';
 export { default as AccordionSummary } from './AccordionSummary';
 export * from './AccordionSummary';
 
+export { default as Alert } from './Alert';
+export * from './Alert';
+
+export { default as AlertTitle } from './AlertTitle';
+export * from './AlertTitle';
+
 export { default as AppBar } from './AppBar';
 export * from './AppBar';
